// src/modules/api/authorize/enums/role.ts
export enum RoleEnum {
    CUSTOMER = "customer",
    ADMIN = "admin",
    SUPER_ADMIN = "super-admin",
}

export enum PermissionName {
    CREATE_USERS = "users.create",
    READ_USERS = "users.read",
    UPDATE_USERS = "users.update",
    DELETE_USERS = "users.delete",
    CREATE_PROVIDER = "provider.create",
    READ_PROVIDER = "provider.read",
    UPDATE_PROVIDER = "provider.update",
    DELETE_PROVIDER = "provider.delete",
    CREATE_TRANSACTIONS = "transactions.create",
    READ_TRANSACTIONS = "transactions.read",
    UPDATE_TRANSACTIONS = "transactions.update",
    DELETE_TRANSACTIONS = "transactions.delete",
    READ_PAYMENT = "payment.read",
    UPDATE_PAYMENT = "payment.update",
    DELETE_PAYMENT = "payment.delete",
    CREATE_DISPUTE = "dispute.create",
    READ_DISPUTE = "dispute.read",
    UPDATE_DISPUTE = "dispute.update",
    DELETE_DISPUTE = "dispute.delete",
    CREATE_REFUND = "refund.create",
    READ_REFUND = "refund.read",
    UPDATE_REFUND = "refund.update",
    DELETE_REFUND = "refund.delete",
    CREATE_REPORT = "report.create",
    READ_REPORT = "report.read",
    UPDATE_REPORT = "report.update",
    DELETE_REPORT = "report.delete",
}